+++ nix/default.nix
@@ -1,6 +1,6 @@
 { sourcesFile ? ./sources.json, system ? builtins.currentSystem }: rec {
   sources = import ./sources.nix { inherit sourcesFile system; };
   plutus = import sources.plutus { };
+  plutus-apps = import sources.plutus-apps { };
   pkgs = plutus.pkgs;
-  pab = import ./pab.nix { inherit plutus; };
 }