+++ plutus-extra/src/Plutus/Contract/Test/Extra.hs
@@ -33,8 +33,18 @@ import Data.List (foldl')
 import Data.Map qualified as Map
 import Data.Maybe (mapMaybe)
 import Data.Row (Row)
-import Data.Text.Prettyprint.Doc (Doc, Pretty, align, colon, indent, pretty, viaShow, vsep, (<+>))
 import Data.Void (Void)
+import Prettyprinter (
+  Doc,
+  Pretty,
+  align,
+  colon,
+  indent,
+  pretty,
+  viaShow,
+  vsep,
+  (<+>),
+ )
 import Prelude
 
 --------------------------------------------------------------------------------